make achievement popup work when api down and user logged in

+++ RetroTrack.Domain/Dtos/UserGameInfoDto.cs
@@ -21,7 +21,7 @@
 
     public class UserAchievement
     {
-        public required int Id { get; set; }
+        public required long Id { get; set; }
         public required string Title { get; set; }
         public required string Description { get; set; }
         public required int Points { get; set; }
